requirements.txt
fastapi
uvicorn
pydantic_settings
websockets
redis[hiredis]